Move ErrorScreen Widget to screen folder
