+++ lib/screens/error/error.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class ErrorScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Error'),
+      ),
+      body: Center(
+        child: Text('ERROR'),
+      ),
+    );
+  }
+}
